Fix: Corrected userInfo access in loadPrivileges thunk
